tools/requirements-python.txt
splat64[mips]>=0.27.2,<1.0.0
tqdm
pycparser

colorama
watchdog
levenshtein
cxxfilt
dataclasses
pyelftools
